requirements.txt
pillow
matplotlib
tqdm